+++ examples/cloudobject.py
@@ -6,21 +6,43 @@ knowing they exact location (bucket, key)
 import pywren_ibm_cloud as pywren
 
 
-def my_function_put(text, internal_storage):
-    co1 = internal_storage.put_object('Temp object test 1: {}'.format(text, ))
-    co2 = internal_storage.put_object('Temp object test 2: {}'.format(text, ))
+def my_function_put(text, ibm_cos):
+    co1 = ibm_cos.put_cobject('Cloudobject test 1: {}'.format(text, ))
+    co2 = ibm_cos.put_cobject('Cloudobject test 2: {}'.format(text, ))
     return [co1, co2]
 
 
-def my_function_get(co, internal_storage):
-    data = internal_storage.get_object(co)
+def my_function_get(co, ibm_cos):
+    data = ibm_cos.get_cobject(co)
     return data
 
 
 if __name__ == "__main__":
+    """
+    Managing cloudobjects with context manager.
+    At the end of the with statement all
+    cloudobjects are automatically deleted.
+    """
+    with pywren.ibm_cf_executor() as pw:
+        pw.call_async(my_function_put, 'Hello Earth')
+        cloudobjects = pw.get_result()
+        pw.map(my_function_get, cloudobjects)
+        print(pw.get_result())
+
+    """
+    Managing cloudobjects without context manager.
+    pw.clean() must be called at the end to delete
+    the cloudobjects created in the same executor as
+    long as you used the default location.
+    Alternatively, you can call pw.clean(cs=cloudobjects)
+    to delete a specific list of cloudobjects.
+    pw.clean(cs=cloudobjects) is mandatory if you created
+    the cloudobjects in a custom location.
+    """
     pw = pywren.ibm_cf_executor()
-    pw.call_async(my_function_put, 'Hello World')
+    pw.call_async(my_function_put, 'Hello Moon')
     cloudobjects = pw.get_result()
     pw.map(my_function_get, cloudobjects)
-    print(pw.get_result())
-    pw.clean()
+    result = pw.get_result()
+    pw.clean()  # or pw.clean(cs=cloudobjects)
+    print(result)