+++ pywren_ibm_cloud/job/job.py
@@ -1,5 +1,8 @@
 import os
+import sys
 import time
+import json
+import textwrap
 import pickle
 import logging
 from pywren_ibm_cloud import utils
@@ -202,3 +205,31 @@ def _create_job(config, internal_storage, executor_id, job_id, func, data, runti
     job_description['metadata'] = host_job_meta
 
     return job_description
+
+
+def clean_job(jobs_to_clean, storage_config, clean_cloudobjects):
+    """
+    Clean the jobs in a separate process
+    """
+    script = """
+    from pywren_ibm_cloud.storage import InternalStorage
+    from pywren_ibm_cloud.storage.utils import clean_bucket
+    from pywren_ibm_cloud.config import JOBS_PREFIX, TEMP_PREFIX
+
+    storage_config = {}
+    jobs_to_clean = {}
+    clean_cloudobjects = {}
+    bucket = storage_config['bucket']
+
+    internal_storage = InternalStorage(storage_config)
+
+    for executor_id, job_id in jobs_to_clean:
+        prefix = '/'.join([JOBS_PREFIX, executor_id, job_id])
+        clean_bucket(bucket, prefix, internal_storage, log=False)
+        if clean_cloudobjects:
+            prefix = '/'.join([TEMP_PREFIX, executor_id, job_id])
+            clean_bucket(bucket, prefix, internal_storage, log=False)
+    """.format(storage_config, jobs_to_clean, clean_cloudobjects)
+
+    cmdstr = '{} -c "{}"'.format(sys.executable, textwrap.dedent(script))
+    os.popen(cmdstr)