+++ pywren_ibm_cloud/storage/storage.py
@@ -3,7 +3,7 @@ import json
 import logging
 import importlib
 from pywren_ibm_cloud.version import __version__
-from pywren_ibm_cloud.config import CACHE_DIR, RUNTIMES_PREFIX, JOBS_PREFIX
+from pywren_ibm_cloud.config import CACHE_DIR, RUNTIMES_PREFIX, JOBS_PREFIX, TEMP_PREFIX
 from pywren_ibm_cloud.utils import is_pywren_function
 from pywren_ibm_cloud.storage.utils import create_status_key, create_output_key, \
     status_key_suffix, init_key_suffix, CloudObject, StorageNoSuchKeyError
@@ -19,6 +19,7 @@ class Storage:
     def __init__(self, pywren_config, storage_backend):
         self.pywren_config = pywren_config
         self.backend = storage_backend
+        self.cloudobject_count = 0
 
         try:
             module_location = 'pywren_ibm_cloud.storage.backends.{}'.format(self.backend)
@@ -35,7 +36,49 @@ class Storage:
         return self.storage_handler
 
     def get_client(self):
-        return self.storage_handler.get_client()
+        client = self.storage_handler.get_client()
+        client.put_cobject = self.put_cobject
+        client.get_cobject = self.get_cobject
+
+        return client
+
+    def put_cobject(self, body, bucket=None, key=None):
+        """
+        Put CloudObject into storage.
+        :param body: data content
+        :param bucket: destination bucket
+        :param key: destination key
+        :return: CloudObject instance
+        """
+        prefix = os.environ.get('PYWREN_EXECUTION_ID', '')
+        name = '{}/cloudobject_{}'.format(prefix, self.cloudobject_count)
+        key = key or '/'.join([TEMP_PREFIX, name])
+        bucket = bucket or self.bucket
+        self.storage_handler.put_object(bucket, key, body)
+        self.cloudobject_count += 1
+
+        return CloudObject(self.backend, bucket, key)
+
+    def get_cobject(self, cloudobject: CloudObject=None, bucket: str=None, key: str=None):
+        """
+        Get CloudObject from storage.
+        :param cloudobject: CloudObject instance
+        :param bucket: destination bucket
+        :param key: destination key
+        :return: body text
+        """
+        if cloudobject:
+            if cloudobject.storage_backend == self.backend:
+                bucket = cloudobject.bucket
+                key = cloudobject.key
+                return self.storage_handler.get_object(bucket, key)
+            else:
+                raise Exception("CloudObject: Invalid Storage backend")
+        elif (bucket and key) or key:
+            bucket = bucket or self.bucket
+            return self.storage_handler.get_object(bucket, key)
+        else:
+            return None
 
 
 class InternalStorage:
@@ -66,6 +109,51 @@ class InternalStorage:
         :return: storage configuration
         """
         return self.config
+
+    def get_client(self):
+        client = self.storage_handler.get_client()
+        client.put_cobject = self.put_cobject
+        client.get_cobject = self.get_cobject
+
+        return client
+
+    def put_cobject(self, body, bucket=None, key=None):
+        """
+        Put CloudObject into storage.
+        :param body: data content
+        :param bucket: destination bucket
+        :param key: destination key
+        :return: CloudObject instance
+        """
+        prefix = os.environ.get('PYWREN_EXECUTION_ID', '')
+        name = '{}/cloudobject_{}'.format(prefix, self.cloudobject_count)
+        key = key or '/'.join([TEMP_PREFIX, name])
+        bucket = bucket or self.bucket
+        self.storage_handler.put_object(bucket, key, body)
+        self.cloudobject_count += 1
+
+        return CloudObject(self.backend, bucket, key)
+
+    def get_cobject(self, cloudobject: CloudObject=None, bucket: str=None, key: str=None):
+        """
+        Get CloudObject from storage.
+        :param cloudobject: CloudObject instance
+        :param bucket: destination bucket
+        :param key: destination key
+        :return: body text
+        """
+        if cloudobject:
+            if cloudobject.storage_backend == self.backend:
+                bucket = cloudobject.bucket
+                key = cloudobject.key
+                return self.storage_handler.get_object(bucket, key)
+            else:
+                raise Exception("CloudObject: Invalid Storage backend")
+        elif (bucket and key) or key:
+            bucket = bucket or self.bucket
+            return self.storage_handler.get_object(bucket, key)
+        else:
+            return None
 
     def put_data(self, key, data):
         """
@@ -100,43 +188,6 @@ class InternalStorage:
         :return: serialized function
         """
         return self.storage_handler.get_object(self.bucket, key)
-
-    def put_object(self, body, bucket=None, key=None):
-        """
-        Put temporal data object into storage.
-        :param key: data key
-        :param data: data content
-        :return: CloudObject instance
-        """
-        prefix = os.environ.get('PYWREN_EXECUTION_ID', '')
-        name = '{}/cloudobject_{}'.format(prefix, self.cloudobject_count)
-        key = key or '/'.join([JOBS_PREFIX, 'cloudobjects', name])
-        bucket = bucket or self.bucket
-        self.storage_handler.put_object(bucket, key, body)
-        self.cloudobject_count += 1
-
-        return CloudObject(self.backend, bucket, key)
-
-    def get_object(self, cloudobject: CloudObject=None, bucket: str=None, key: str=None):
-        """
-        get temporal data object from storage.
-        :param cloudobject: CloudObject instance
-        :param key: data bucket
-        :param key: data key
-        :return: body text
-        """
-        if cloudobject:
-            if cloudobject.storage_backend == self.backend:
-                bucket = cloudobject.bucket
-                key = cloudobject.key
-                return self.storage_handler.get_object(bucket, key)
-            else:
-                raise Exception("CloudObject: Invalid Storage backend")
-        elif (bucket and key) or key:
-            bucket = bucket or self.bucket
-            return self.storage_handler.get_object(bucket, key)
-        else:
-            return None
 
     def get_job_status(self, executor_id, job_id):
         """