+++ pywren_ibm_cloud/tests.py
@@ -152,10 +152,10 @@ class TestMethods:
         return counter
 
     @staticmethod
-    def my_map_function_ibm_cos(key_i, bucket_name, ibm_cos):
+    def my_map_function_storage(key_i, bucket_name, storage):
         print('I am processing the object /{}/{}'.format(bucket_name, key_i))
         counter = {}
-        data = ibm_cos.get_object(Bucket=bucket_name, Key=key_i)['Body'].read()
+        data = storage.get_object(Bucket=bucket_name, Key=key_i)['Body'].read()
         for line in data.splitlines():
             for word in line.decode('utf-8').split():
                 if word not in counter:
@@ -173,14 +173,14 @@ class TestMethods:
         return final_result
 
     @staticmethod
-    def my_cloudobject_put(obj, internal_storage):
+    def my_cloudobject_put(obj, storage):
         counter = TestMethods.my_map_function_obj(obj, 0)
-        cloudobject = internal_storage.put_object(pickle.dumps(counter))
+        cloudobject = storage.put_cobject(pickle.dumps(counter))
         return cloudobject
 
     @staticmethod
-    def my_cloudobject_get(results, internal_storage):
-        data = [pickle.loads(internal_storage.get_object(cloudobject)) for cloudobject in results]
+    def my_cloudobject_get(results, storage):
+        data = [pickle.loads(storage.get_cobject(cloudobject)) for cloudobject in results]
         return TestMethods.my_reduce_function(data)
 
 
@@ -344,16 +344,17 @@ class TestPywren(unittest.TestCase):
         self.assertEqual(result, self.__class__.cos_result_to_compare)
 
     def test_storage_handler(self):
-        print('Testing ibm_cos function arg...')
+        print('Testing "storage" function arg...')
         iterdata = [[key, STORAGE_CONFIG['bucket']] for key in TestUtils.list_test_keys()]
         pw = pywren.function_executor(config=CONFIG)
-        pw.map_reduce(TestMethods.my_map_function_ibm_cos, iterdata, TestMethods.my_reduce_function)
+        pw.map_reduce(TestMethods.my_map_function_storage, iterdata, TestMethods.my_reduce_function)
         result = pw.get_result()
         self.assertEqual(result, self.__class__.cos_result_to_compare)
 
     def test_chunks_bucket(self):
         print('Testing chunks on a bucket...')
-        data_prefix = STORAGE_CONFIG['bucket'] + '/' + PREFIX + '/'
+        sb = STORAGE_CONFIG['backend']
+        data_prefix = sb + '://' + STORAGE_CONFIG['bucket'] + '/' + PREFIX + '/'
 
         pw = pywren.function_executor(config=CONFIG)
         futures = pw.map_reduce(TestMethods.my_map_function_obj, data_prefix, TestMethods.my_reduce_function,
@@ -370,7 +371,8 @@ class TestPywren(unittest.TestCase):
 
     def test_chunks_bucket_one_reducer_per_object(self):
         print('Testing chunks on a bucket with one reducer per object...')
-        data_prefix = STORAGE_CONFIG['bucket'] + '/' + PREFIX + '/'
+        sb = STORAGE_CONFIG['backend']
+        data_prefix = sb + '://' + STORAGE_CONFIG['bucket'] + '/' + PREFIX + '/'
 
         pw = pywren.function_executor(config=CONFIG)
         futures = pw.map_reduce(TestMethods.my_map_function_obj, data_prefix, TestMethods.my_reduce_function,
@@ -388,11 +390,12 @@ class TestPywren(unittest.TestCase):
 
     def test_cloudobject(self):
         print('Testing cloudobjects...')
-        data_prefix = STORAGE_CONFIG['bucket'] + '/' + PREFIX + '/'
-        pw = pywren.function_executor(config=CONFIG)
-        pw.map_reduce(TestMethods.my_cloudobject_put, data_prefix, TestMethods.my_cloudobject_get)
-        result = pw.get_result()
-        self.assertEqual(result, self.__class__.cos_result_to_compare)
+        sb = STORAGE_CONFIG['backend']
+        data_prefix = sb + '://' + STORAGE_CONFIG['bucket'] + '/' + PREFIX + '/'
+        with pywren.ibm_cf_executor(config=CONFIG) as pw:
+            pw.map_reduce(TestMethods.my_cloudobject_put, data_prefix, TestMethods.my_cloudobject_get)
+            result = pw.get_result()
+            self.assertEqual(result, self.__class__.cos_result_to_compare)
 
 
 def print_help():