+++ pywren_ibm_cloud/utils.py
@@ -186,12 +186,11 @@ def b64str_to_bytes(str_data):
 
 
 def split_object_url(obj_url):
-    if '://' in obj_url:
-        sb, path = obj_url.split('://')
-    else:
-        sb = 'ibm_cos'
-        path = obj_url
+    if '://' not in obj_url:
+        raise Exception('Object url must contain the storage backend prefix, '
+                        'for example: cos://')
 
+    sb, path = obj_url.split('://')
     sb = 'ibm_cos' if sb == 'cos' else sb
 
     bucket, full_key = path.split('/', 1) if '/' in path else (path, '')
@@ -264,7 +263,7 @@ def verify_args(func, iterdata, extra_params):
     data = format_data(iterdata, extra_params)
 
     # Verify parameters
-    non_verify_args = ['ibm_cos', 'swift', 'internal_storage', 'id', 'rabbitmq']
+    non_verify_args = ['ibm_cos', 'swift', 'storage', 'id', 'rabbitmq']
     func_sig = inspect.signature(func)
 
     new_parameters = list()